requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
typing_extensions==4.12.2
Werkzeug==3.1.3
jupyterlab==4.3.5
pdfplumber==0.11.5
fuzzywuzzy==0.18.0
spacy==3.8.4
# After installing spacy, run the following command to download the language model:
# python -m spacy download en_core_web_sm
pdf2image==1.17.0
pytesseract==0.3.13
pytest==8.3.4
pytest-mock==3.14.0
python-Levenshtein==0.26.1
